Split to lib/bin crates, Use "midir", not PortMidi
Starting to flesh this out into an actual application, starting by
making the midi input its own module. For this structuring, PortMidi
didn't work well, Midir seems to be a better alternative. It now works
through a threaded model with a messagebus.

Old example hardcoded actions are still included in the binary package,
but more configurable stuff is coming up.