Simplify MIDI message parsing
Doesn't need an additional couple of crates.

Now matches against binary values directly in the parse method instead
of having an intermediary enum to go from number to message types.

Given that the raw binary number is on the same line as the enum value
it results in, the binary numbers are self-documenting enough.