The Great Config Parsing Commit
Implements parsing Yaml to the intermediary "RawConfig" data format
which is no more than a share representation of primitive data types
that can be represented in yaml, or perhaps json.

Yaml is parsed into RawConfig.

Then we read the top-level "version" field from the RawConfig hash, and
with that, get a config processor. Since this is the first version,
there is only one version of config so far ("1"), which is covered by
version1_processor.rs

version1_processor turns all this RawConfig structure into all the data
types that have been built so far; scopes, actions, macros, event
matchers, ...