Implemented getting focused window info w/o shell
Instead of running commands through `std::process::Command` to retrieve
this data, use libxdo and libx to access it directly.

This removes the reliance on the xdotool* and xprop programs being
installed and available in PATH.

[*] xdotool still needs to be installed for the bindings to libxdo

Also changes the focus adapter module name from xdo to x11, since it's
not purely xdo.

Closes #5.