Implement control actions in task_main
In the binary, all 3 control actions are now implemented. This required
moving a couple of things around to have the path of the config file
around etc. Restarting task_main altogether is most straightforward from
a data accessing perspective by having task_main return a bool
indicating whether it is to be restarted, since main.rs hangs on to the
data that was used to launch it.