Add variables module, parse var notation to tokens
Not ready for actual use yet. This parses variable notation (that should
be found between % characters) into a set of tokens that can further be
used to retrieve data.

The function notation parsing is currently a large function and should
be broken up further into its own parsing component.

A further refactor will be needed to parse these variable notation
strings as part of the string embedding them, so it can produce
structures that contain markers to where in the string replacement is to
be done.

It also still needs to add support for default values.